Generalize prohibition on unreleased code.
In addition to adding links to the project website, prohibit any action
which might encourage the public to use unreleased materials -- e.g.
publishing to various package management systems, recommending direct
SCM checkouts in documentation...

Furthermore, change "download and use" to "download or use" because it is
problematic if non-developers do either.  By switching to "or", the
language expands to encompass not only nightly builds and such but also
source control, which is not usually referred to as a "download".