Clarify canonical distribution requirement.
Add explicit language requiring that official releases be uploaded to
the canonical distribution channel.

Eliminate incorrect language suggesting that a release is not "released"
unless it is uploaded.  If, say, a release vote passes and then the
artifacts are made available to the general public through alternate
channels, the release remains an "act of the foundation".